requirements.txt
annotated-types
anyio
asttokens
accelerate
beautifulsoup4
biopython
certifi
charset-normalizer
colorama
comm
contourpy
cycler
datasets
debugpy
decorator
distro
dotenv
exceptiongroup
executing
fonttools
h11
httpcore
httpx
idna
importlib_metadata
ipykernel
ipython
jedi
jiter
joblib
jupyter_client
jupyter_core
kiwisolver
loguru
matplotlib
matplotlib-inline
nest_asyncio
numpy
openai
packaging
pandas
parso
pickleshare
pillow
platformdirs
prompt_toolkit
psutil
pure_eval
pydantic
pydantic_core
pynvml
Pygments
pyparsing
python-dateutil
python-dotenv
requests
scikit-learn
scipy
seaborn
setuptools
six
sniffio
soupsieve
stack-data
timm
thop
threadpoolctl
tornado
torch
torchvision
transformers
tqdm
traitlets
typing-inspection
typing_extensions
triton
tzdata
urllib3
wandb
wcwidth
wheel
zipp